requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
openai